rename Components name: ...-template => ...-components
