+++ src/components/trip-days-item-component.js
