+++ src/components/trip-days-list-component.js
