+++ src/components/trip-event-item-component.js
