+++ src/components/trip-events-list-component.js
